+++ test/specs/controller.bubble.tests.js
@@ -1,6 +1,10 @@
 describe('Chart.controllers.bubble', function() {
 	describe('auto', jasmine.fixture.specs('controller.bubble'));
 
+	it('should be registered as dataset controller', function() {
+		expect(typeof Chart.controllers.bubble).toBe('function');
+	});
+
 	it('should be constructed', function() {
 		var chart = window.acquireChart({
 			type: 'bubble',
