+++ test/specs/controller.line.tests.js
@@ -1,6 +1,10 @@
 describe('Chart.controllers.line', function() {
 	describe('auto', jasmine.fixture.specs('controller.line'));
 
+	it('should be registered as dataset controller', function() {
+		expect(typeof Chart.controllers.line).toBe('function');
+	});
+
 	it('should be constructed', function() {
 		var chart = window.acquireChart({
 			type: 'line',
